fix: address Nezha proxy review feedback
