+++ crates/rginx-config/src/validate/tests.rs
@@ -12,6 +12,7 @@ use super::{DEFAULT_GRPC_HEALTH_CHECK_PATH, validate};
 mod listeners;
 mod route;
 mod runtime;
+mod server_client_ip;
 mod server_http3;
 mod server_limits;
 mod server_tls;
